backend/requirements.txt
# Crime Risk Prediction API - Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Geospatial
h3==3.7.6
shapely==2.0.2

# Data Science
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# HTTP & APIs
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

